pyproject.toml
[project]
name = "autosubtitles"
version = "0.1.0"
description = "generate subtitles for youtube videos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp[default]==2025.6.30",
    "mcp==1.10.1",
    "mcp-server==0.1.4",
    "python-dotenv==1.0.0",
    "python-magic==0.4.27",
    "openai-whisper",
    "ffmpeg-python==0.2.0",
    "pydub==0.25.1",
    "fastapi==0.115.12",
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-core==0.3.59",
    "langchain-experimental==0.3.4",
    "langchain-openai==0.3.5",
    "langgraph==0.3.31",
    "langchain-mcp-adapters==0.1.8",
]

[tool.uv.sources]
openai-whisper = { path = "whisper", editable = true }
